scraper_v2/requirements.txt
# Facebook Scraper v2.0 Requirements
# ===================================

# Core Dependencies
selenium>=4.10.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL

# API Server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# ML & NLP
textblob>=0.17.0
nltk>=3.8.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Firebase (optional)
firebase-admin>=6.0.0

# Google Sheets (optional)
gspread>=5.10.0
oauth2client>=4.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ===================================
# Optional: Advanced ML with BERT
# Uncomment if you want advanced sentiment
# ===================================
# torch>=2.0.0
# transformers>=4.30.0

# ===================================
# Installation:
#   pip install -r requirements.txt
#
# For advanced ML:
#   pip install torch transformers
# ===================================
